Add files via upload
Steganography is the practice of hiding information in other data to conceal the existence of the hidden message. This technique has been used throughout history to transmit secret information covertly and has been adapted for use in the digital age. In digital steganography, the hidden message is embedded in an image, audio file, or other digital media in a way that is not easily noticeable. The hidden message can be recovered by extracting the information from the cover media, typically using specialized software or tools.
Steganography is often used in security and privacy protection to conceal sensitive information from prying eyes. For example, a person may want to hide a confidential message in an image file and send the image file over the internet with the hope that the message will be secret.